refactor: add required room dependency to core
Room is required for Core so that the Entity and PrimaryKey annotations can be added to the models
