+++ core/src/main/java/com/chesire/malime/core/models/SeriesModel.kt
@@ -3,10 +3,10 @@ package com.chesire.malime.core.models
 import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.chesire.malime.server.flags.SeriesStatus
-import com.chesire.malime.server.flags.SeriesType
-import com.chesire.malime.server.flags.Subtype
-import com.chesire.malime.server.flags.UserSeriesStatus
+import com.chesire.malime.core.flags.SeriesStatus
+import com.chesire.malime.core.flags.SeriesType
+import com.chesire.malime.core.flags.Subtype
+import com.chesire.malime.core.flags.UserSeriesStatus
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
