+++ core/src/main/java/com/chesire/malime/core/models/UserModel.kt
@@ -2,7 +2,7 @@ package com.chesire.malime.core.models
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.chesire.malime.server.flags.Service
+import com.chesire.malime.core.flags.Service
 
 @Entity
 data class UserModel(
